requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
pydantic==2.11.3
pydantic-settings==2.9.1
mysqlclient>=2.2
email-validator==2.0.0 